pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cen"
version = "0.2.0"
description = "CEN AI Concierge — Community Equity Navigators AOP/DAG Platform"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "networkx>=3.2",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "structlog>=24.1",
]

[project.optional-dependencies]
llm = ["llama-cpp-python>=0.2"]
privacy = ["presidio-analyzer>=2.2", "presidio-anonymizer>=2.2"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "anyio>=4.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cen"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
